Extracted scan parsing
