+++ .vscode/c_cpp_properties.json
@@ -4,9 +4,14 @@
             "name": "Mac",
             "includePath": [
                 "${default}",
-                "/usr/local/include/sdlwrapper",
+                "${workspaceFolder}/include",
+                "${workspaceFolder}/common",
+                "/usr/include/SDL2",
                 "/usr/include/eigen3",
-                "/usr/local/include",
+                "/usr/local/include/sdlwrapper",
+                "/usr/local/include/cmdapp",
+                "/usr/local/include/config",
+                "/usr/local/include/simple_test",
                 "/Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/lib/clang/15.0.0/include/"
             ],
             "defines": [],
@@ -18,6 +23,7 @@
             "includePath": [
                 "${default}",
                 "${workspaceFolder}/include",
+                "${workspaceFolder}/common",
                 "/usr/include/SDL2",
                 "/usr/include/eigen3",
                 "/usr/local/include/sdlwrapper",
