+++ .vscode/launch.json
@@ -13,8 +13,6 @@
                 "ex_data/scan1/second.conf",
                 "--method",
                 "vanilla",
-                // "--gui",
-                "--bench"
             ],
             "stopAtEntry": false,
             "cwd": "${workspaceFolder}",
